Remove 12 no-op scalar casts restating a variable's own declared type
Sites where the cast type equals the sole repo-wide declaration of the
operand and no pointer or aggregate declaration of that name exists.
Eight TUs verified byte-identical; four candidates (fuelCell,
NW_levcontr, Andross, track_dolphin) moved bytes and were reverted -
those casts sit on narrow locals where MWCC emits a real truncation.

+++ src/dlls/objects/604/604.c
@@ -279,7 +279,7 @@ void snowclaw_updateMountAttack(GameObject* obj, GameObject* mount)
                                             magnitude);
     }
 
-    if (ObjAnim_AdvanceCurrentMove((int)obj, moveStep, (f32)(u8)framesThisStep,
+    if (ObjAnim_AdvanceCurrentMove((int)obj, moveStep, (f32)framesThisStep,
                                                                     NULL) != 0 &&
         (obj)->anim.currentMove != inner->moveIdBase)
     {
